requirements.txt
aiogram==2.24
aiogram-dialog==1.9.0
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
exceptiongroup==1.1.0
fake-useragent==1.1.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.1
multidict==6.0.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.2
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
yarl==1.8.2
